Added login page
This update adds the login page, allowing users to authenticate using Firebase. Functionality for redirecting to the main page after successfully logging in has been implemented. Additionally, animations have been included to enhance the user experience when interacting with the login panels. Contact links and a reference to the website have been provided for further information and support.